CLOUDSTACK-8459: Authn requests must use either HTTP POST/Artifact binding
As per SAML v2.0 spec section 4.1.2, the SP provider should create authn requests using
either HTTP POST or HTTP Artifact binding to transfer the message through a
user agent (browser in our case). The use of HTTP Redirect was one of the reasons
why this plugin failed to work for some IdP servers that enforce this.

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>